tests/wpt/web-platform-tests/css/css-transforms/parsing/rotate-parsing-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Transform Module Level 2: parsing rotate with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-transforms-2/#propdef-rotate">
<meta name="assert" content="rotate supports the full grammar 'none | <number>{3}? <angle>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("rotate", "none");

test_valid_value("rotate", "0deg");

test_valid_value("rotate", "100 200 300 400grad");
test_valid_value("rotate", "400grad 100 200 300", "100 200 300 400grad");

test_valid_value("rotate", "x 400grad", "1 0 0 400grad");
test_valid_value("rotate", "400grad x", "1 0 0 400grad");

test_valid_value("rotate", "y 400grad", "0 1 0 400grad");
test_valid_value("rotate", "400grad y", "0 1 0 400grad");

test_valid_value("rotate", "z 400grad", "0 0 1 400grad");
test_valid_value("rotate", "400grad z", "0 0 1 400grad");
</script>
</body>
</html>
